[svn] - forgot one long value

+++ ggz-client-libs/ggzmod/ggzmod.c
@@ -4,7 +4,7 @@
  * Project: ggzmod
  * Date: 10/14/01
  * Desc: GGZ game module functions
- * $Id: ggzmod.c 6149 2004-07-17 20:41:02Z josef $
+ * $Id: ggzmod.c 6150 2004-07-17 20:42:55Z josef $
  *
  * This file contains the backend for the ggzmod library.  This
  * library facilitates the communication between the GGZ server (ggz)
@@ -1099,7 +1099,7 @@ int ggzmod_spectator_get_ranking(GGZSpectatorSeat *seat, int *ranking)
 	return 0;
 }
 
-int ggzmod_spectator_get_highscore(GGZSpectatorSeat *seat, long *highscore)
+int ggzmod_spectator_get_highscore(GGZSpectatorSeat *seat, int *highscore)
 {
 	return 0;
 }
